Refactoring, should cherry-pick to separate PR
