+++ Source/VCGeneration/CommandTransformations.cs
@@ -1,0 +1,17 @@
+using Microsoft.Boogie;
+using VC;
+
+namespace VCGeneration;
+
+public static class CommandTransformations
+{
+  public static Cmd AssertIntoAssume(VCGenOptions options, Cmd c)
+  {
+    if (c is AssertCmd assertCmd)
+    {
+      return VCGen.AssertTurnedIntoAssume(options, assertCmd);
+    }
+
+    return c;
+  }
+}
